% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Calculate the Cube of a Number}
\usage{
cube(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
The cube of x
}
\description{
Calculate the Cube of a Number
}
\examples{
cube(3)
cube(2)  # Returns 8
cube(4)  # Returns 64
}
