% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Calculate the Square of a Number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
The square of x
}
\description{
Calculate the Square of a Number
}
\examples{
square(5)
square(5)  # Returns 25
square(3)  # Returns 9
}
